requirements.txt
# Luzmo API client
requests>=2.31.0

# Anthropic Claude API
anthropic>=0.39.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# Configuration
pyyaml>=6.0
